requirements.txt
Pillow
requests
Flask
Flask-Cors
google-cloud-storage